pyproject.toml
[project]
name = "hr-analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fairlearn>=0.13.0",
    "fastapi>=0.136.0",
    "google-genai>=1.73.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "optuna>=4.8.0",
    "pandas>=3.0.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "shap>=0.51.0",
    "uvicorn>=0.46.0",
    "xgboost>=3.2.0",
]
